pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "credit-card-fraud-detection"
version = "1.0.0"
description = "End-to-end MLOps pipeline for credit card fraud detection using XGBoost"
requires-python = ">=3.11"

dependencies = [
    # data & ML
    "pandas",
    "numpy",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "imbalanced-learn>=0.12",

    # data validation
    "pandera>=0.19",

    # experiment tracking
    "mlflow>=2.12",

    # data versioning
    "dvc>=3.49",

    # API serving
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",

    # orchestration
    "dagster>=1.7",
    "dagster-webserver>=1.7",

    # monitoring
    "evidently>=0.4",

    # config
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
